requirements.txt
# Deep Learning - versions réelles
tensorflow==2.16.2
keras==2.16.0  # Cette version n'existe pas, on la retire

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.14

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Utilities
matplotlib==3.7.2
pillow==10.0.0



